requirements.txt
django==4.1.6
djangorestframework==3.13.1
pillow==9.4.0
python-decouple==3.7
postgres==4.0
psycopg2-binary==2.9.5
psycopg2==2.9.5
django-seed==0.3.1